@@ -1,28 +1,34 @@
 source 'https://rubygems.org'
+git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-gem 'rails', '~>5.0.0'
+ruby '2.7.0'
+
+# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
+gem 'rails', '~> 6.0.3', '>= 6.0.3.2'
 gem 'chef', '~> 12'
-gem 'jbuilder', '~> 1.2'
+# Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
+gem 'jbuilder', '~> 2.7'
 gem 'jquery-rails'
 gem 'jwt' # For Zendesk SSO
-gem 'rails_config', '~> 0.4.2'
+gem 'config', '~> 2.2', '>= 2.2.1'  # Replacement of rails_config gem
 gem 'rb-readline', '~> 0.5.2', require: false
 gem 'sass-rails', '>= 4.0.3'
-gem 'turbolinks', '~> 2.2.1'
-gem 'unicorn-rails', '~> 1.1.0'
+gem 'turbolinks', '~> 5'
+gem 'unicorn-rails', '~> 2.2', '>= 2.2.1'
 gem 'omniauth', '~> 1.3.2'
-gem 'nokogiri', '1.10.4'
-gem 'pg', '~> 0.15' # active_record 4.2.8 pins this but doesn't manifest this in the gemspec for some reason
+gem 'nokogiri', '1.10.10'
+gem 'pg', '>= 0.18', '< 2.0' # active_record 4.2.8 pins this but doesn't manifest this in the gemspec for some reason
 gem 'mixlib-authentication', '~> 1.3'
 gem 'sentry-raven'
-gem 'responders', '~> 2.1.2'
+gem 'responders', '~> 3.0', '>= 3.0.1'
 gem 'newrelic_rpm'
-gem 'doorkeeper', '~> 4.0'
-gem 'sprockets', '~> 4.0'
+gem 'doorkeeper', '~> 5.4'
+gem "sprockets-rails", git: 'git://github.com/rails/sprockets-rails.git'
 
+gem 'foundation-rails', git: 'https://github.com/foundation/foundation-rails.git', tag: 'v5.5.3.2'
 gem 'veil', git: 'https://github.com/chef/chef_secrets'
-gem 'omniauth-chef', git: 'https://github.com/chef/omniauth-chef', tag: 'v0.3.0'
-gem 'chef-web-core', git: 'https://github.com/chef/chef-web-core.git'
+gem 'omniauth-chef', git: 'https://github.com/chef/omniauth-chef'
+# gem 'chef-web-core', git: 'https://github.com/chef/chef-web-core.git'
 
 #
 # These gems require a javascript runtime.  We don't want to ship a
@@ -30,16 +36,16 @@ gem 'chef-web-core', git: 'https://github.com/chef/chef-web-core.git'
 # don't get automatically required by rake.
 #
 group :node do
-  gem 'coffee-rails', '~> 4.2'
-  gem 'uglifier', '~> 2.4.0'
+  gem 'coffee-rails', '~> 5.0'
+  gem 'uglifier', '~> 4.2'
 end
 
 group :development, :test do
   # Loading it this way makes it work with Ruby 2.1.2. See
   # https://github.com/nixme/jazz_hands/issues/25
-  # jazz_hands no support for Rails 5 onwards 
   # gem 'jazz_hands', git: 'https://github.com/chef/jazz_hands', branch: 'praj/bring-your-own-debugger'
-  gem 'rspec-rails', '~> 3.5'
+  gem 'rspec-rails', '~> 4.0', '>= 4.0.1'
+  gem 'rails-controller-testing'
   gem 'pry-byebug'
   gem 'mailcatcher'
 end
@@ -47,11 +53,10 @@ end
 group :development do
   gem 'better_errors'
   gem 'binding_of_caller'
-  # replaced by sprockets
-  #gem 'quiet_assets'
+  # gem 'quiet_assets'
   gem 'spring' # App preloading
   gem 'spring-commands-rspec'
-  gem 'thor', '~> 0.18.0'
+  gem 'thor', '~> 1.0', '>= 1.0.1'
 end
 
 group :doc do
@@ -59,7 +64,7 @@ group :doc do
 end
 
 group :test do
-  gem 'capybara', '~> 2.4.4'
+  gem 'capybara', '~> 3.33'
   gem 'factory_girl_rails', '~> 4.4.0'
   gem 'selenium-webdriver', '~> 2.53.0'
   gem 'timecop'
