@@ -1,4 +1,4 @@
-class ChangeColumnTypeForResourceOwnerId < ActiveRecord::Migration
+class ChangeColumnTypeForResourceOwnerId < ActiveRecord::Migration[6.0]
   def change
 
     change_table :oauth_access_grants do |t|
