@@ -1,4 +1,4 @@
-OcId::Application.routes.draw do
+Rails.application.routes.draw do
   root 'home#index'
   get 'id', to: 'home#index'
 
